calendar and spring cleaning
I changed up the css for the navbar a bit, I changed the color schema for the news page and fixed a slight wrapping error on the news page. I fixed a sizing issue on the flyers page, and I changed its look slightly. I also changed up the calendar so that it can resize, and made it so that a user can go backwards or forwards in the months